error reporting and remove phone validation
