+++ plugins/UCRM-Client-Signup/src/assets/ucrm-client-signup-form-be3bda63c1dd28be4ae04cb1304c6c78.js
@@ -1,0 +1,23 @@
+"use strict"
+define("ucrm-client-signup-form/app",["exports","ucrm-client-signup-form/resolver","ember-load-initializers","ucrm-client-signup-form/config/environment"],function(e,t,n,r){Object.defineProperty(e,"__esModule",{value:!0})
+var o=Ember.Application.extend({modulePrefix:r.default.modulePrefix,podModulePrefix:r.default.podModulePrefix,Resolver:t.default});(0,n.default)(o,r.default.modulePrefix),e.default=o}),define("ucrm-client-signup-form/components/bs-accordion",["exports","ember-bootstrap/components/bs-accordion"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-accordion/item",["exports","ember-bootstrap/components/bs-accordion/item"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-accordion/item/body",["exports","ember-bootstrap/components/bs-accordion/item/body"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-accordion/item/title",["exports","ember-bootstrap/components/bs-accordion/item/title"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-alert",["exports","ember-bootstrap/components/bs-alert"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-button-group",["exports","ember-bootstrap/components/bs-button-group"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-button-group/button",["exports","ember-bootstrap/components/bs-button-group/button"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-button",["exports","ember-bootstrap/components/bs-button"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-carousel",["exports","ember-bootstrap/components/bs-carousel"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-carousel/slide",["exports","ember-bootstrap/components/bs-carousel/slide"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-collapse",["exports","ember-bootstrap/components/bs-collapse"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown",["exports","ember-bootstrap/components/bs-dropdown"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/button",["exports","ember-bootstrap/components/bs-dropdown/button"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/menu",["exports","ember-bootstrap/components/bs-dropdown/menu"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/menu/divider",["exports","ember-bootstrap/components/bs-dropdown/menu/divider"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/menu/item",["exports","ember-bootstrap/components/bs-dropdown/menu/item"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/menu/link-to",["exports","ember-bootstrap/components/bs-dropdown/menu/link-to"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-dropdown/toggle",["exports","ember-bootstrap/components/bs-dropdown/toggle"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form",["exports","ember-bootstrap-cp-validations/components/bs-form"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element",["exports","ember-bootstrap-cp-validations/components/bs-form/element"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/control",["exports","ember-bootstrap/components/bs-form/element/control"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/control/checkbox",["exports","ember-bootstrap/components/bs-form/element/control/checkbox"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/control/input",["exports","ember-bootstrap/components/bs-form/element/control/input"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/control/textarea",["exports","ember-bootstrap/components/bs-form/element/control/textarea"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/errors",["exports","ember-bootstrap/components/bs-form/element/errors"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/feedback-icon",["exports","ember-bootstrap/components/bs-form/element/feedback-icon"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/help-text",["exports","ember-bootstrap/components/bs-form/element/help-text"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/label",["exports","ember-bootstrap/components/bs-form/element/label"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/layout/horizontal",["exports","ember-bootstrap/components/bs-form/element/layout/horizontal"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})
+define("ucrm-client-signup-form/components/bs-form/element/layout/horizontal/checkbox",["exports","ember-bootstrap/components/bs-form/element/layout/horizontal/checkbox"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/layout/inline",["exports","ember-bootstrap/components/bs-form/element/layout/inline"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/layout/inline/checkbox",["exports","ember-bootstrap/components/bs-form/element/layout/inline/checkbox"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/layout/vertical",["exports","ember-bootstrap/components/bs-form/element/layout/vertical"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/element/layout/vertical/checkbox",["exports","ember-bootstrap/components/bs-form/element/layout/vertical/checkbox"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-form/group",["exports","ember-bootstrap/components/bs-form/group"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal-simple",["exports","ember-bootstrap/components/bs-modal-simple"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal",["exports","ember-bootstrap/components/bs-modal"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/body",["exports","ember-bootstrap/components/bs-modal/body"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/dialog",["exports","ember-bootstrap/components/bs-modal/dialog"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/footer",["exports","ember-bootstrap/components/bs-modal/footer"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/header",["exports","ember-bootstrap/components/bs-modal/header"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/header/close",["exports","ember-bootstrap/components/bs-modal/header/close"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-modal/header/title",["exports","ember-bootstrap/components/bs-modal/header/title"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-nav",["exports","ember-bootstrap/components/bs-nav"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-nav/item",["exports","ember-bootstrap/components/bs-nav/item"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-nav/link-to",["exports","ember-bootstrap/components/bs-nav/link-to"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-navbar",["exports","ember-bootstrap/components/bs-navbar"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-navbar/content",["exports","ember-bootstrap/components/bs-navbar/content"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-navbar/link-to",["exports","ember-bootstrap/components/bs-navbar/link-to"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-navbar/nav",["exports","ember-bootstrap/components/bs-navbar/nav"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-navbar/toggle",["exports","ember-bootstrap/components/bs-navbar/toggle"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-popover",["exports","ember-bootstrap/components/bs-popover"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-popover/element",["exports","ember-bootstrap/components/bs-popover/element"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-progress",["exports","ember-bootstrap/components/bs-progress"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-progress/bar",["exports","ember-bootstrap/components/bs-progress/bar"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-tab",["exports","ember-bootstrap/components/bs-tab"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-tab/pane",["exports","ember-bootstrap/components/bs-tab/pane"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-tooltip",["exports","ember-bootstrap/components/bs-tooltip"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/bs-tooltip/element",["exports","ember-bootstrap/components/bs-tooltip/element"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})
+define("ucrm-client-signup-form/components/ember-popper-targeting-parent",["exports","ember-popper/components/ember-popper-targeting-parent"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/ember-popper",["exports","ember-popper/components/ember-popper"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/components/user-details",["exports","ucrm-client-signup-form/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({ajax:Ember.inject.service(),store:Ember.inject.service(),classNames:["container-fluid"],processing:!1,actions:{submit:function(e){var n=this
+e.validate().then(function(e){var r=e.validations
+n.set("pending",!0),n.set("processing",!0),n.set("failure",!1),r.get("isValid")&&(n.get("ajax").post(t.default.APP.host,{headers:{"Content-Type":"application/json"},data:{pluginAppKey:t.default.APP.pluginAppKey,client:{organizationId:1,clientType:1,firstName:n.get("model.client.firstName"),lastName:n.get("model.client.lastName"),street1:n.get("model.client.street1"),street2:n.get("model.client.street2"),city:n.get("model.client.city"),countryId:n.get("model.client.countryId"),stateId:n.get("model.client.stateId"),zipCode:n.get("model.client.zipCode"),username:n.get("model.client.email"),contacts:[{email:n.get("model.client.email"),phone:n.get("model.client.phone"),name:n.get("model.client.firstName")+" "+n.get("model.client.lastName")}]}}}).catch(function(e){void 0!==e.payload&&null!==e.payload&&(!0===e.payload.redirect?(n.set("failure",!1),n.get("changeRoute")("signup.account")):n.set("errors",e.payload.errors)),n.set("pending",!1),n.set("failure",!0)}).then(function(){!0!==n.get("failure")&&n.get("changeRoute")("signup.complete"),n.set("pending",!1)}),n.set("processing",!1))})}}})}),define("ucrm-client-signup-form/controllers/signup/account",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Controller.extend({actions:{changeRoute:function(e){this.transitionToRoute(e)}}})}),define("ucrm-client-signup-form/helpers/app-version",["exports","ucrm-client-signup-form/config/environment","ember-cli-app-version/utils/regexp"],function(e,t,n){function r(e){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},o=t.default.APP.version,i=r.versionOnly||r.hideSha,u=r.shaOnly||r.hideVersion,l=null
+return i&&(r.showExtended&&(l=o.match(n.versionExtendedRegExp)),l||(l=o.match(n.versionRegExp))),u&&(l=o.match(n.shaRegExp)),l?l[0]:o}Object.defineProperty(e,"__esModule",{value:!0}),e.appVersion=r,e.default=Ember.Helper.helper(r)}),define("ucrm-client-signup-form/helpers/bs-contains",["exports","ember-bootstrap/helpers/bs-contains"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"bsContains",{enumerable:!0,get:function(){return t.bsContains}})}),define("ucrm-client-signup-form/helpers/bs-eq",["exports","ember-bootstrap/helpers/bs-eq"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"eq",{enumerable:!0,get:function(){return t.eq}})}),define("ucrm-client-signup-form/helpers/cancel-all",["exports","ember-concurrency/helpers/cancel-all"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"cancelAll",{enumerable:!0,get:function(){return t.cancelAll}})}),define("ucrm-client-signup-form/helpers/is-value",["exports"],function(e){function t(e){return e[0]===e[1]}Object.defineProperty(e,"__esModule",{value:!0}),e.isValue=t,e.default=Ember.Helper.helper(t)}),define("ucrm-client-signup-form/helpers/perform",["exports","ember-concurrency/helpers/perform"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"perform",{enumerable:!0,get:function(){return t.perform}})}),define("ucrm-client-signup-form/helpers/pluralize",["exports","ember-inflector/lib/helpers/pluralize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("ucrm-client-signup-form/helpers/singularize",["exports","ember-inflector/lib/helpers/singularize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("ucrm-client-signup-form/helpers/task",["exports","ember-concurrency/helpers/task"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"task",{enumerable:!0,get:function(){return t.task}})}),define("ucrm-client-signup-form/initializers/app-version",["exports","ember-cli-app-version/initializer-factory","ucrm-client-signup-form/config/environment"],function(e,t,n){Object.defineProperty(e,"__esModule",{value:!0})
+var r=void 0,o=void 0
+n.default.APP&&(r=n.default.APP.name,o=n.default.APP.version),e.default={name:"App Version",initialize:(0,t.default)(r,o)}}),define("ucrm-client-signup-form/initializers/container-debug-adapter",["exports","ember-resolver/resolvers/classic/container-debug-adapter"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"container-debug-adapter",initialize:function(){var e=arguments[1]||arguments[0]
+e.register("container-debug-adapter:main",t.default),e.inject("container-debug-adapter:main","namespace","application:main")}}}),define("ucrm-client-signup-form/initializers/data-adapter",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"data-adapter",before:"store",initialize:function(){}}}),define("ucrm-client-signup-form/initializers/ember-concurrency",["exports","ember-concurrency/initializers/ember-concurrency"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"initialize",{enumerable:!0,get:function(){return t.initialize}})}),define("ucrm-client-signup-form/initializers/ember-data",["exports","ember-data/setup-container","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("ucrm-client-signup-form/initializers/export-application-global",["exports","ucrm-client-signup-form/config/environment"],function(e,t){function n(){var e=arguments[1]||arguments[0]
+if(!1!==t.default.exportApplicationGlobal){var n
+if("undefined"!=typeof window)n=window
+else if("undefined"!=typeof global)n=global
+else{if("undefined"==typeof self)return
+n=self}var r,o=t.default.exportApplicationGlobal
+r="string"==typeof o?o:Ember.String.classify(t.default.modulePrefix),n[r]||(n[r]=e,e.reopen({willDestroy:function(){this._super.apply(this,arguments),delete n[r]}}))}}Object.defineProperty(e,"__esModule",{value:!0}),e.initialize=n,e.default={name:"export-application-global",initialize:n}}),define("ucrm-client-signup-form/initializers/injectStore",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"injectStore",before:"store",initialize:function(){}}}),define("ucrm-client-signup-form/initializers/load-bootstrap-config",["exports","ucrm-client-signup-form/config/environment","ember-bootstrap/config"],function(e,t,n){function r(){n.default.load(t.default["ember-bootstrap"]||{})}Object.defineProperty(e,"__esModule",{value:!0}),e.initialize=r,e.default={name:"load-bootstrap-config",initialize:r}}),define("ucrm-client-signup-form/initializers/store",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"store",after:"ember-data",initialize:function(){}}}),define("ucrm-client-signup-form/initializers/transforms",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"transforms",before:"store",initialize:function(){}}}),define("ucrm-client-signup-form/instance-initializers/ember-data",["exports","ember-data/initialize-store-service"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("ucrm-client-signup-form/models/client",["exports","ember-data","ember-cp-validations"],function(e,t,n){Object.defineProperty(e,"__esModule",{value:!0})
+var r=t.default.attr,o=(0,n.buildValidations)({firstName:[(0,n.validator)("presence",!0),(0,n.validator)("length",{min:3})],lastName:[(0,n.validator)("presence",!0),(0,n.validator)("length",{min:3})],email:[(0,n.validator)("presence",!0),(0,n.validator)("format",{type:"email"})],phone:[(0,n.validator)("presence",!0)]})
+e.default=t.default.Model.extend(o,{firstName:r("string"),lastName:r("string"),contacts:r(),street1:r("string"),zipCode:r("string"),city:r("string"),email:r("string"),phone:r("string"),stateId:r("number",{defaultValue:6}),countryId:r("number",{defaultValue:249}),agreedToTAC:r("boolean",{defaultValue:!1}),fullAddress:Ember.computed("street1","zipCode","city",function(){return this.get("street1")+", "+this.get("city")+", CO. "+this.get("zipCode")})})}),define("ucrm-client-signup-form/resolver",["exports","ember-resolver"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("ucrm-client-signup-form/router",["exports","ucrm-client-signup-form/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0})
+var n=Ember.Router.extend({location:t.default.locationType,rootURL:t.default.rootURL})
+n.map(function(){this.route("signup",{path:"/"},function(){this.route("account",{path:"/"}),this.route("complete")})}),e.default=n}),define("ucrm-client-signup-form/routes/application",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({})}),define("ucrm-client-signup-form/routes/signup",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({ajax:Ember.inject.service(),model:function(){return Ember.RSVP.hash({client:this.get("store").createRecord("client")})}})}),define("ucrm-client-signup-form/routes/signup/account",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({queryParams:{expired:{refreshModel:!0}},model:function(e){return!0===e.expired&&this.set("controller.expired",!0),Ember.RSVP.hash({client:this.modelFor("signup").client})}})})
+define("ucrm-client-signup-form/routes/signup/complete",["exports","ucrm-client-signup-form/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({model:function(){return t.default.APP.completionText}})}),define("ucrm-client-signup-form/services/ajax",["exports","ember-ajax/services/ajax"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/templates/application",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"18+zaeeG",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","application-wrapper"],[7],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n"],[8]],"hasEval":false}',meta:{moduleName:"ucrm-client-signup-form/templates/application.hbs"}})}),define("ucrm-client-signup-form/templates/components/ember-popper-targeting-parent",["exports","ember-popper/templates/components/ember-popper-targeting-parent"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/templates/components/ember-popper",["exports","ember-popper/templates/components/ember-popper"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/templates/components/user-details",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"hMibXInM",block:'{"symbols":["form","form","key","value"],"statements":[[4,"bs-form",null,[["model","novalidate","onSubmit"],[[20,["model","client"]],true,[25,"action",[[19,0,[]],"submit",[20,["model","client"]]],null]]],{"statements":[[0,"  "],[6,"div"],[9,"class","form-row"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-md-6 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["text","firstName","given-name","First Name","firstName",true]]],false],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-md-6 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["text","lastName","family-name","Last Name","lastName",true]]],false],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n  "],[6,"div"],[9,"class","form-row"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-md-6 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["email","email","email","Email","email",true]]],false],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-md-6 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["phone","phone","tel tel-national","Phone Number","phone",true]]],false],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","form-row"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-md-6 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["text","street1","street-address","Street Address","street1",true]]],false],[0,"\\n    "],[8],[0,"\\n"],[0,"    "],[6,"div"],[9,"class","col-md-3 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["text","city","address-level2","City","city",true]]],false],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-md-3 mb-2"],[7],[0,"\\n      "],[1,[25,"component",[[19,1,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["text","zip","postal-code","Zip","zipCode",true]]],false],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n"],[0,"\\n"],[4,"unless",[[20,["error"]]],null,{"statements":[[4,"if",[[20,["pending"]]],null,{"statements":[[0,"      "],[6,"div"],[9,"class","alert alert-primary"],[9,"role","alert"],[7],[0,"Submitting Information"],[8],[0,"\\n      "],[6,"div"],[9,"class","row justify-content-center"],[7],[0,"\\n        "],[6,"div"],[9,"class","col-auto"],[7],[0,"\\n          "],[6,"img"],[9,"src","https://s3.amazonaws.com/shared-charuwts/images/spinner.gif"],[9,"style","display:inline-block;"],[7],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[6,"br"],[9,"clear","all"],[7],[8],[0,"\\n"]],"parameters":[]},{"statements":[[4,"if",[[20,["failure"]]],null,{"statements":[[0,"        "],[6,"div"],[9,"class","alert alert-danger"],[9,"role","alert"],[7],[0,"Sorry, there was an error with your request."],[8],[0,"\\n"],[4,"if",[[20,["errors"]]],null,{"statements":[[4,"bs-form",null,[["class","model","novalidate"],["container-fluid",[20,["model","client"]],true]],{"statements":[[0,"            "],[6,"div"],[9,"class","row"],[7],[0,"\\n"],[4,"each",[[25,"-each-in",[[20,["errors"]]],null]],null,{"statements":[[0,"                "],[6,"div"],[9,"class","col-6"],[7],[0,"\\n"],[4,"if",[[25,"is-value",[[19,3,[]],"username"],null]],null,{"statements":[[0,"                    "],[6,"span"],[9,"style","text-transform:capitalize;"],[7],[0,"email"],[8],[0," - "],[1,[19,4,[]],false],[0,"\\n                    "],[1,[25,"component",[[19,2,["element"]]],[["controlType","id","autocomplete","placeholder","property","required"],["email","email","email","Email","email",true]]],false],[0,"\\n"]],"parameters":[]},{"statements":[[0,"                    "],[6,"span"],[9,"style","text-transform:capitalize;"],[7],[1,[19,3,[]],false],[8],[0," - "],[1,[19,4,[]],false],[0,"\\n                    "],[1,[25,"component",[[19,2,["element"]]],[["placeholder","property","required"],[[19,3,[]],[19,3,[]],true]]],false],[0,"\\n"]],"parameters":[]}],[0,"                "],[8],[0,"\\n"]],"parameters":[3,4]},null],[0,"            "],[8],[0,"\\n"]],"parameters":[2]},null]],"parameters":[]},null],[0,"\\n"]],"parameters":[]},null]],"parameters":[]}]],"parameters":[]},null],[0,"\\n\\n  "],[6,"div"],[9,"class","row"],[7],[0,"\\n    "],[6,"div"],[9,"class","col text-center"],[7],[0,"\\n"],[4,"if",[[20,["processing"]]],null,{"statements":[[0,"        "],[6,"div"],[9,"class","btn btn-dark disabled"],[9,"id","proceed"],[7],[6,"span"],[7],[0,"Processing"],[8],[8],[0,"\\n"]],"parameters":[]},{"statements":[[4,"if",[[25,"get",[[20,["model","client","validations"]],"isInvalid"],null]],null,{"statements":[[0,"        "],[6,"div"],[9,"class","btn btn-dark disabled"],[9,"id","proceed"],[7],[6,"span"],[7],[0,"Provide Info"],[8],[8],[0,"\\n"]],"parameters":[]},{"statements":[[0,"        "],[6,"button"],[9,"class","btn btn-secondary"],[9,"id","proceed"],[9,"type","submit"],[7],[6,"span"],[7],[0,"Signup"],[8],[8],[0,"\\n      "]],"parameters":[]}]],"parameters":[]}],[0,"    "],[8],[0,"\\n  "],[8],[0,"\\n"]],"parameters":[1]},null]],"hasEval":false}',meta:{moduleName:"ucrm-client-signup-form/templates/components/user-details.hbs"}})}),define("ucrm-client-signup-form/templates/signup",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"F/Wynq23",block:'{"symbols":[],"statements":[[0,"\\n"],[0,"\\n"],[6,"div"],[9,"class","signup-wrapper"],[7],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"ucrm-client-signup-form/templates/signup.hbs"}})}),define("ucrm-client-signup-form/templates/signup/account",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"VwMivuel",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","account-wrapper"],[7],[0,"\\n  "],[1,[25,"user-details",null,[["model","changeRoute"],[[20,["model"]],[25,"action",[[19,0,[]],"changeRoute"],null]]]],false],[0,"\\n"],[8]],"hasEval":false}',meta:{moduleName:"ucrm-client-signup-form/templates/signup/account.hbs"}})}),define("ucrm-client-signup-form/templates/signup/complete",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"/QhJmBQ/",block:'{"symbols":[],"statements":[[6,"h1"],[9,"class","text-center"],[7],[0,"Signup Complete"],[8],[0,"\\n"],[6,"div"],[9,"class","text-center"],[7],[0,"\\n  "],[1,[18,"model"],false],[0,"\\n"],[8],[0,"\\n"],[1,[18,"outlet"],false]],"hasEval":false}',meta:{moduleName:"ucrm-client-signup-form/templates/signup/complete.hbs"}})}),define("ucrm-client-signup-form/validators/alias",["exports","ember-cp-validations/validators/alias"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/belongs-to",["exports","ember-cp-validations/validators/belongs-to"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/collection",["exports","ember-cp-validations/validators/collection"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/confirmation",["exports","ember-cp-validations/validators/confirmation"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/date",["exports","ember-cp-validations/validators/date"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/dependent",["exports","ember-cp-validations/validators/dependent"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/ds-error",["exports","ember-cp-validations/validators/ds-error"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/exclusion",["exports","ember-cp-validations/validators/exclusion"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/format",["exports","ember-cp-validations/validators/format"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/has-many",["exports","ember-cp-validations/validators/has-many"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/inclusion",["exports","ember-cp-validations/validators/inclusion"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/inline",["exports","ember-cp-validations/validators/inline"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/length",["exports","ember-cp-validations/validators/length"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/messages",["exports","ember-cp-validations/validators/messages"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/number",["exports","ember-cp-validations/validators/number"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/validators/presence",["exports","ember-cp-validations/validators/presence"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("ucrm-client-signup-form/config/environment",[],function(){try{var e="ucrm-client-signup-form/config/environment",t=document.querySelector('meta[name="'+e+'"]').getAttribute("content"),n={default:JSON.parse(unescape(t))}
+return Object.defineProperty(n,"__esModule",{value:!0}),n}catch(t){throw new Error('Could not read config from meta tag with name "'+e+'".')}}),runningTests||require("ucrm-client-signup-form/app").default.create({rootElement:"#ember-signup",host:"http://localhost:8080/_plugins/ucrm-client-signup/public.php",completionText:"completiontextinformation",pluginAppKey:"null",name:"ucrm-client-signup-form",version:"1.0.0+9a7beb79"})
