+++ plugins/UCRM-Client-Signup/src/manifest.json
@@ -5,7 +5,7 @@
     "displayName": "UCRM Client Signup",
     "description": "A UCRM plugin that provides a frontend form for clients to signup.",
     "url": "https://www.charuwts.com/plugins",
-    "version": "1.0.0",
+    "version": "1.0.1",
     "ucrmVersionCompliancy": {
       "min": "2.11.1",
       "max": null
